test(gateway): consolidate update coverage
