--- src/handlers/gateway/gateway.update.fixture.test.tsx
@@ -1,426 +1,0 @@
-import { expect, test } from "bun:test";
-import { mkdirSync, readFileSync, writeFileSync } from "node:fs";
-import { join } from "node:path";
-import {
-  CreateGatewayCommand,
-  CreateGatewayRuleCommand,
-  CreateGatewayTargetCommand,
-  DeleteGatewayCommand,
-  DeleteGatewayRuleCommand,
-  DeleteGatewayTargetCommand,
-  GetGatewayCommand,
-  GetGatewayRuleCommand,
-  GetGatewayTargetCommand,
-} from "@aws-sdk/client-bedrock-agentcore-control";
-import {
-  CreateRoleCommand,
-  DeleteRoleCommand,
-  DeleteRolePolicyCommand,
-  PutRolePolicyCommand,
-} from "@aws-sdk/client-iam";
-import { CoreClient } from "../../core";
-import { createControlClient, createIamClient } from "../../core/factories";
-import {
-  createSilentLogger,
-  fixtureFactories,
-  isRecording,
-  matchGolden,
-  TestGlobalConfigAccessor,
-  testIO,
-} from "../../testing";
-import { createRootHandler } from "../index";
-
-const REGION = "us-east-1";
-const FIXTURES = join(import.meta.dir, "__fixtures__", "update");
-const RESOURCE_STATE = join(FIXTURES, "resources.json");
-const GATEWAY_NAME = "agentcore-cli-gateway-update-fixture";
-const ROLE_NAME = "AgentCoreCliGatewayUpdateFixtureRole";
-const POLICY_NAME = "AgentCoreCliGatewayUpdateFixture";
-const HTTP_TARGET_NAME = "http-update-fixture";
-const CONNECTOR_TARGET_NAME = "web-search-update-fixture";
-const FLOW_TIMEOUT = 600_000;
-
-type FixtureState = {
-  gatewayId: string;
-  gatewayArn: string;
-  targetId: string;
-  connectorId: string;
-  ruleId: string;
-};
-
-function createFixtureCore(): CoreClient {
-  const { createControlClient, createDataClient, createIamClient } = fixtureFactories(FIXTURES);
-  return new CoreClient({
-    createControlClient,
-    createDataClient,
-    createIamClient,
-    logger: createSilentLogger(),
-  });
-}
-
-async function run(args: string[]): Promise<string> {
-  const io = testIO();
-  const root = createRootHandler(createFixtureCore(), {
-    io: io.io,
-    logger: createSilentLogger(),
-    globalConfigAccessor: new TestGlobalConfigAccessor(),
-  });
-  await root.route(["node", "agentcore", ...args, "--region", REGION]);
-  return io.stdout();
-}
-
-async function waitUntil<T>(
-  operation: () => Promise<T>,
-  done: (response: T) => boolean,
-): Promise<T> {
-  for (let attempt = 0; attempt < 60; attempt += 1) {
-    const response = await operation();
-    if (done(response)) return response;
-    await Bun.sleep(2_000);
-  }
-  throw new Error("Timed out waiting for fixture resource state");
-}
-
-async function ignoreMissing(operation: () => Promise<unknown>): Promise<void> {
-  try {
-    await operation();
-  } catch (error) {
-    if (!["ResourceNotFoundException", "NoSuchEntityException"].includes((error as Error).name)) {
-      throw error;
-    }
-  }
-}
-
-async function waitUntilMissing(operation: () => Promise<unknown>): Promise<void> {
-  for (let attempt = 0; attempt < 60; attempt += 1) {
-    try {
-      await operation();
-    } catch (error) {
-      if ((error as Error).name === "ResourceNotFoundException") return;
-      throw error;
-    }
-    await Bun.sleep(2_000);
-  }
-  throw new Error("Timed out waiting for fixture resource deletion");
-}
-
-// Prerequisites use direct clients so recording Update never rewrites Create-owned fixtures.
-async function setup(): Promise<FixtureState> {
-  const control = createControlClient({ region: REGION });
-  const iam = createIamClient({ region: REGION });
-  const role = await iam.send(
-    new CreateRoleCommand({
-      RoleName: ROLE_NAME,
-      AssumeRolePolicyDocument: JSON.stringify({
-        Version: "2012-10-17",
-        Statement: [
-          {
-            Effect: "Allow",
-            Principal: { Service: "bedrock-agentcore.amazonaws.com" },
-            Action: "sts:AssumeRole",
-          },
-        ],
-      }),
-    }),
-  );
-  if (!role.Role?.Arn) throw new Error("IAM did not return the fixture role ARN");
-
-  const gateway = await control.send(
-    new CreateGatewayCommand({
-      name: GATEWAY_NAME,
-      roleArn: role.Role.Arn,
-      authorizerType: "NONE",
-      description: "Gateway before update",
-    }),
-  );
-  if (!gateway.gatewayId || !gateway.gatewayArn) {
-    throw new Error("CreateGateway did not return fixture identifiers");
-  }
-  await waitUntil(
-    () => control.send(new GetGatewayCommand({ gatewayIdentifier: gateway.gatewayId })),
-    (response) => response.status === "READY",
-  );
-
-  await iam.send(
-    new PutRolePolicyCommand({
-      RoleName: ROLE_NAME,
-      PolicyName: POLICY_NAME,
-      PolicyDocument: JSON.stringify({
-        Version: "2012-10-17",
-        Statement: [
-          {
-            Effect: "Allow",
-            Action: "bedrock-agentcore:InvokeGateway",
-            Resource: gateway.gatewayArn,
-          },
-          {
-            Effect: "Allow",
-            Action: "bedrock-agentcore:InvokeWebSearch",
-            Resource: "arn:aws:bedrock-agentcore:us-east-1:aws:tool/web-search.v1",
-          },
-        ],
-      }),
-    }),
-  );
-
-  const target = await control.send(
-    new CreateGatewayTargetCommand({
-      gatewayIdentifier: gateway.gatewayId,
-      name: HTTP_TARGET_NAME,
-      description: "Target before update",
-      targetConfiguration: {
-        http: {
-          passthrough: {
-            endpoint: "https://example.com",
-            protocolType: "CUSTOM",
-          },
-        },
-      },
-    }),
-  );
-  const connector = await control.send(
-    new CreateGatewayTargetCommand({
-      gatewayIdentifier: gateway.gatewayId,
-      name: CONNECTOR_TARGET_NAME,
-      description: "Connector before update",
-      targetConfiguration: {
-        mcp: {
-          connector: {
-            source: { connectorId: "web-search" },
-            configurations: [
-              {
-                name: "WebSearch",
-                parameterValues: { maxResults: 10 },
-              },
-            ],
-          },
-        },
-      },
-      credentialProviderConfigurations: [{ credentialProviderType: "GATEWAY_IAM_ROLE" }],
-    }),
-  );
-  if (!target.targetId || !connector.targetId) {
-    throw new Error("CreateGatewayTarget did not return fixture identifiers");
-  }
-  await Promise.all(
-    [target.targetId, connector.targetId].map((targetId) =>
-      waitUntil(
-        () =>
-          control.send(
-            new GetGatewayTargetCommand({
-              gatewayIdentifier: gateway.gatewayId,
-              targetId,
-            }),
-          ),
-        (response) => response.status === "READY",
-      ),
-    ),
-  );
-
-  const rule = await control.send(
-    new CreateGatewayRuleCommand({
-      gatewayIdentifier: gateway.gatewayId,
-      priority: 10,
-      actions: [
-        {
-          routeToTarget: {
-            staticRoute: {
-              targetName: HTTP_TARGET_NAME,
-            },
-          },
-        },
-      ],
-      description: "Rule before update",
-    }),
-  );
-  if (!rule.ruleId) throw new Error("CreateGatewayRule did not return the fixture rule ID");
-  await waitUntil(
-    () =>
-      control.send(
-        new GetGatewayRuleCommand({
-          gatewayIdentifier: gateway.gatewayId,
-          ruleId: rule.ruleId,
-        }),
-      ),
-    (response) => response.status === "ACTIVE",
-  );
-
-  const state = {
-    gatewayId: gateway.gatewayId,
-    gatewayArn: gateway.gatewayArn,
-    targetId: target.targetId,
-    connectorId: connector.targetId,
-    ruleId: rule.ruleId,
-  };
-  mkdirSync(FIXTURES, { recursive: true });
-  writeFileSync(RESOURCE_STATE, `${JSON.stringify(state, null, 2)}\n`);
-  return state;
-}
-
-async function cleanup(state: FixtureState): Promise<void> {
-  const control = createControlClient({ region: REGION });
-  const iam = createIamClient({ region: REGION });
-  await ignoreMissing(() =>
-    control.send(
-      new DeleteGatewayRuleCommand({
-        gatewayIdentifier: state.gatewayId,
-        ruleId: state.ruleId,
-      }),
-    ),
-  );
-  await waitUntilMissing(() =>
-    control.send(
-      new GetGatewayRuleCommand({
-        gatewayIdentifier: state.gatewayId,
-        ruleId: state.ruleId,
-      }),
-    ),
-  );
-  for (const targetId of [state.targetId, state.connectorId]) {
-    await ignoreMissing(() =>
-      control.send(
-        new DeleteGatewayTargetCommand({
-          gatewayIdentifier: state.gatewayId,
-          targetId,
-        }),
-      ),
-    );
-    await waitUntilMissing(() =>
-      control.send(
-        new GetGatewayTargetCommand({
-          gatewayIdentifier: state.gatewayId,
-          targetId,
-        }),
-      ),
-    );
-  }
-  await ignoreMissing(() =>
-    control.send(new DeleteGatewayCommand({ gatewayIdentifier: state.gatewayId })),
-  );
-  await waitUntilMissing(() =>
-    control.send(new GetGatewayCommand({ gatewayIdentifier: state.gatewayId })),
-  );
-  await ignoreMissing(() =>
-    iam.send(new DeleteRolePolicyCommand({ RoleName: ROLE_NAME, PolicyName: POLICY_NAME })),
-  );
-  await ignoreMissing(() => iam.send(new DeleteRoleCommand({ RoleName: ROLE_NAME })));
-}
-
-async function verify<T>(
-  operation: () => Promise<T>,
-  done: (response: T) => boolean,
-): Promise<void> {
-  if (!isRecording()) return;
-  await waitUntil(operation, done);
-}
-
-test(
-  "updates a Gateway, Target, Connector, and Rule through the real Core",
-  async () => {
-    const state = isRecording()
-      ? await setup()
-      : (JSON.parse(readFileSync(RESOURCE_STATE, "utf8")) as FixtureState);
-    const control = createControlClient({ region: REGION });
-
-    try {
-      const gatewayStdout = await run([
-        "gateway",
-        "update",
-        "--id",
-        state.gatewayId,
-        "--description",
-        "Gateway after update",
-      ]);
-      matchGolden(FIXTURES, "gateway-update.golden.json", gatewayStdout);
-      expect(JSON.parse(gatewayStdout).description).toBe("Gateway after update");
-      await verify(
-        () => control.send(new GetGatewayCommand({ gatewayIdentifier: state.gatewayId })),
-        (response) =>
-          response.status === "READY" && response.description === "Gateway after update",
-      );
-
-      const targetStdout = await run([
-        "gateway",
-        "target",
-        "update",
-        "--gateway-id",
-        state.gatewayId,
-        "--target-id",
-        state.targetId,
-        "--description",
-        "Target after update",
-      ]);
-      matchGolden(FIXTURES, "target-update.golden.json", targetStdout);
-      expect(JSON.parse(targetStdout).description).toBe("Target after update");
-      await verify(
-        () =>
-          control.send(
-            new GetGatewayTargetCommand({
-              gatewayIdentifier: state.gatewayId,
-              targetId: state.targetId,
-            }),
-          ),
-        (response) => response.status === "READY" && response.description === "Target after update",
-      );
-
-      const connectorStdout = await run([
-        "gateway",
-        "connector",
-        "update",
-        "--gateway-id",
-        state.gatewayId,
-        "--id",
-        state.connectorId,
-        "--description",
-        "Connector after update",
-      ]);
-      matchGolden(FIXTURES, "connector-update.golden.json", connectorStdout);
-      expect(JSON.parse(connectorStdout).description).toBe("Connector after update");
-      await verify(
-        () =>
-          control.send(
-            new GetGatewayTargetCommand({
-              gatewayIdentifier: state.gatewayId,
-              targetId: state.connectorId,
-            }),
-          ),
-        (response) =>
-          response.status === "READY" && response.description === "Connector after update",
-      );
-
-      const ruleStdout = await run([
-        "gateway",
-        "rule",
-        "update",
-        "--gateway-id",
-        state.gatewayId,
-        "--rule-id",
-        state.ruleId,
-        "--priority",
-        "20",
-        "--description",
-        "Rule after update",
-      ]);
-      matchGolden(FIXTURES, "rule-update.golden.json", ruleStdout);
-      const updatedRule = JSON.parse(ruleStdout);
-      expect(updatedRule.priority).toBe(20);
-      expect(updatedRule.description).toBe("Rule after update");
-      await verify(
-        () =>
-          control.send(
-            new GetGatewayRuleCommand({
-              gatewayIdentifier: state.gatewayId,
-              ruleId: state.ruleId,
-            }),
-          ),
-        (response) =>
-          response.status === "ACTIVE" &&
-          response.priority === 20 &&
-          response.description === "Rule after update",
-      );
-    } finally {
-      if (isRecording()) await cleanup(state);
-    }
-  },
-  FLOW_TIMEOUT,
-);
